backend/auth-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email-validator]==2.9.0
email-validator>=2.0.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
python-multipart==0.0.12
